Add files via upload
This is Proteon the Masked One from the discord server.  Made a few replacements for the turtletouched theme started by JustEthanIGuess and others.  Update includes 11 expressions and a proposed 'expressioned' neck to correspond with the head, if such a thing is possible.  Photoshop doc included with the three layers to each folder: Base_Untinted, texture, and Base_Skin.